+++ backend/http/router.go
@@ -107,7 +107,7 @@ func StartHttp(Service ImgService, storage *storage.Storage, cache FileCache) {
 
 	// Public routes
 	api.HandleFunc("GET /public/publicUser", publicUserGetHandler)
-	api.HandleFunc("GET /public/dl", withHashFile(publicDlHandler))
+	api.HandleFunc("GET /public/dl", withHashFile(rawHandler))
 	api.HandleFunc("GET /public/share", withHashFile(publicShareHandler))
 
 	// Settings routes